feat: refactor form ui
